--- src/features/BlogPreviewBody/presentations/index.module.scss
@@ -1,1 +1,0 @@
-@use '@/styles/variables' as *;
