--- src/features/BlogPreviewUsage/presentations/index.module.scss
@@ -1,48 +1,0 @@
-@use "@/styles/variables" as *;
-
-.container {
-  .title {
-    font-size: $text-2xl;
-    font-weight: bold;
-    margin-bottom: 16px;
-
-    @media (max-width: $mobile) {
-      font-size: $text-sm;
-      margin-bottom: 9px;
-    }
-  }
-
-  .text {
-    font-size: $text-base;
-    margin-bottom: 16px;
-
-    @media (max-width: $mobile) {
-      display: none;
-    }
-  }
-
-  .mt24 {
-    margin-top: 24px;
-  }
-
-  .inputArea {
-    .input {
-      border-radius: $rounded;
-      border: 1px solid #cbcbcb;
-      background-color: $white;
-      padding: 8px;
-      width: 700px;
-
-      @media (max-width: $mobile) {
-        max-width: 70vw;
-      }
-    }
-
-    .inputButton {
-      border-radius: $rounded;
-      border: 1px solid #cbcbcb;
-      padding: 8px;
-      background-color: $white;
-    }
-  }
-}
