+++ src/screens/Preview/index.tsx
@@ -2,13 +2,15 @@ import React from "react";
 import { BlogPreview } from "@/features/BlogPreview";
 import styles from "./index.module.scss";
 
-type Props = {};
+type Props = {
+  notionId?: string;
+};
 
-export const PreviewScreen: React.FC<Props> = (props) => {
+export const PreviewScreen: React.FC<Props> = ({ notionId }) => {
   return (
     <main className={styles.wrapper}>
       <h1 className={styles.pageTitle}>ブログプレビュー</h1>
-      <BlogPreview />
+      <BlogPreview id={notionId} />
     </main>
   );
 };
